requirements.txt
Flask==2.3.3
pymongo==4.4.0
gunicorn==20.1.0
python-dotenv==1.0.0
